+++ wn/_queries.py
@@ -280,10 +280,11 @@ def find_proposed_ilis(
 
 
 def find_entries(
-        id: str = None,
-        forms: Sequence[str] = None,
-        pos: str = None,
-        lexicon_rowids: Sequence[int] = None,
+    id: str = None,
+    forms: Sequence[str] = None,
+    pos: str = None,
+    lexicon_rowids: Sequence[int] = None,
+    normalized: bool = False,
 ) -> Iterator[_Word]:
     conn = connect()
     cte = ''
@@ -294,12 +295,12 @@ def find_entries(
         params.append(id)
     if forms:
         cte = f'WITH wordforms(s) AS (VALUES {_vs(forms)})'
-        conditions.append('''
+        or_norm = 'OR normalized_form IN wordforms' if normalized else ''
+        conditions.append(f'''
             e.rowid IN
                (SELECT entry_rowid
                   FROM forms
-                 WHERE form IN wordforms
-                    OR normalized_form IN wordforms)
+                 WHERE form IN wordforms {or_norm})
         '''.strip())
         params.extend(forms)
     if pos:
@@ -338,6 +339,7 @@ def find_senses(
     forms: Sequence[str] = None,
     pos: str = None,
     lexicon_rowids: Sequence[int] = None,
+    normalized: bool = False,
 ) -> Iterator[_Sense]:
     conn = connect()
     cte = ''
@@ -348,12 +350,12 @@ def find_senses(
         params.append(id)
     if forms:
         cte = f'WITH wordforms(s) AS (VALUES {_vs(forms)})'
-        conditions.append('''
+        or_norm = 'OR normalized_form IN wordforms' if normalized else ''
+        conditions.append(f'''
             s.entry_rowid IN
                (SELECT entry_rowid
                   FROM forms
-                 WHERE form IN wordforms
-                    OR normalized_form IN wordforms)
+                 WHERE form IN wordforms {or_norm})
         '''.strip())
         params.extend(forms)
     if pos:
@@ -386,6 +388,7 @@ def find_synsets(
     pos: str = None,
     ili: str = None,
     lexicon_rowids: Sequence[int] = None,
+    normalized: bool = False,
 ) -> Iterator[_Synset]:
     conn = connect()
     cte = ''
@@ -398,12 +401,12 @@ def find_synsets(
         params.append(id)
     if forms:
         cte = f'WITH wordforms(s) AS (VALUES {_vs(forms)})'
-        join = '''\
+        or_norm = 'OR normalized_form IN wordforms' if normalized else ''
+        join = f'''\
           JOIN (SELECT _s.entry_rowid, _s.synset_rowid, _s.entry_rank
                   FROM forms AS f
                   JOIN senses AS _s ON _s.entry_rowid = f.entry_rowid
-                 WHERE f.form IN wordforms
-                    OR f.normalized_form IN wordforms) AS s
+                 WHERE f.form IN wordforms {or_norm}) AS s
             ON s.synset_rowid = ss.rowid
         '''.strip()
         params.extend(forms)